# doc/swagger.yaml
swagger: "2.0"
info:
  description: "This is the FlopBox swagger."
  version: "1.0.0"
  title: "Swagger FlopBox"
host: "localhost:8080"
basePath: "/flopbox"
...
schemes:
- "http"
paths:
...
  /client/{flopBoxUsername}:
    post:
      tags:
        - "client"
      summary: "Creates a folder or upload a file/folder to a FTP server"
      operationId: "addToFTPServer"
      parameters:
      - in: "header"
        name: "username"
        description: "FTP username"
        required: true
        type: string
      - in: "header"
        name: "password"
        description: "FTP password"
        required: true
        type: string
      - in: "header"
        name: "flopBoxPassword"
        description: "FlopBox login password"
        required: true
        type: string
      - in: "path"
        name: "flopBoxUsername"
        description: "FlopBox login username"
        required: true
        type: string
      - in: "query"
        name: "alias"
        description: "FTP alias"
        required: true
        type: "string"
      - in: "query"
        name: "path"
        description: "Final location where the file or folder will be"
        required: true
        type: string
      - in: "query"
        name: "directoryName"
        description: "Name of the directory to create"
        required: false
        type: string
      - in: "formData"
        name: "file"
        description: "File to upload"
        required: false
        type: file
      consumes: 
      - 'multipart/form-data'
      produces:
      - 'text/plain'
      responses:
        "200":
          description: "Folder created | file/folder uploaded successfuly"
        "400":
          description: "Unable to create folder | Unable to send file/folder"
        "401":
          description: "FlopBox bad authentication"
        "409":
          description: "Param error | Alias doesn't exists"
    get:
      tags:
        - "client"
      summary: "Lists files of a FTP server"
      operationId: "listFTPServer"
      parameters:
      - in: "header"
        name: "username"
        description: "FTP username"
        required: true
        type: string
      - in: "header"
        name: "password"
        description: "FTP password"
        required: true
        type: string
      - in: "header"
        name: "flopBoxPassword"
        description: "FlopBox login password"
        required: true
        type: string
      - in: "path"
        name: "flopBoxUsername"
        description: "FlopBox login username"
        required: true
        type: string
      - in: "query"
        name: "alias"
        description: "FTP alias"
        required: true
        type: "string"
      - in: "query"
        name: "path"
        description: "Final location where the file or folder will be"
        required: true
        type: string
      produces:
      - 'application/json'
      responses:
        "200":
          description: "Listed successfuly"
        "401":
          description: "FlopBox bad authentication"
        "409":
          description: "Param error | Alias doesn't exists"
    put:
      tags:
        - "client"
      summary: "Renames file/folder of a FTP server"
      operationId: "renameFTPServer"
      parameters:
      - in: "header"
        name: "username"
        description: "FTP username"
        required: true
        type: string
      - in: "header"
        name: "password"
        description: "FTP password"
        required: true
        type: string
      - in: "header"
        name: "flopBoxPassword"
        description: "FlopBox login password"
        required: true
        type: string
      - in: "path"
        name: "flopBoxUsername"
        description: "FlopBox login username"
        required: true
        type: string
      - in: "query"
        name: "alias"
        description: "FTP alias"
        required: false
        type: "string"
      - in: "query"
        name: "path"
        description: "File/folder location"
        required: false
        type: string
      - in: "query"
        name: "newName"
        description: "New name for the file/folder"
        required: false
        type: string
      - in: "query"
        name: "mode"
        description: "Mode selected (passive or active)"
        required: false
        type: string
      produces:
      - 'text/plain'
      responses:
        "200":
          description: "Renamed | mode changed successfuly"
        "400":
          description: "Unable to rename file/folder | incorrect param"
        "401":
          description: "FlopBox bad authentication"
        "409":
          description: "Param error | Alias doesn't exists"
        "500":
          description: "File/folder doesn't exists"
    delete:
      tags:
        - "client"
      summary: "Removes file/folder of a FTP server"
      operationId: "removeFTPServer"
      parameters:
      - in: "header"
        name: "username"
        description: "FTP username"
        required: true
        type: string
      - in: "header"
        name: "password"
        description: "FTP password"
        required: true
        type: string
      - in: "header"
        name: "flopBoxPassword"
        description: "FlopBox login password"
        required: true
        type: string
      - in: "path"
        name: "flopBoxUsername"
        description: "FlopBox login username"
        required: true
        type: string
      - in: "query"
        name: "alias"
        description: "FTP alias"
        required: true
        type: "string"
      - in: "query"
        name: "path"
        description: "File/folder location"
        required: true
        type: string
      produces:
      - 'text/plain'
      responses:
        "200":
          description: "Removed successfuly"
        "400":
          description: "Unable to remove file/folder"
        "401":
          description: "FlopBox bad authentication"
        "409":
          description: "Param error | Alias doesn't exists"
  /client/{flopBoxUsername}/download:
    get:
      tags:
        - "client"
      summary: "Retrieves file/folder of a FTP server"
      operationId: "retrieveFTPServer"
      parameters:
      - in: "header"
        name: "username"
        description: "FTP username"
        required: true
        type: string
      - in: "header"
        name: "password"
        description: "FTP password"
        required: true
        type: string
      - in: "header"
        name: "flopBoxPassword"
        description: "FlopBox login password"
        required: true
        type: string
      - in: "path"
        name: "flopBoxUsername"
        description: "FlopBox login username"
        required: true
        type: string
      - in: "query"
        name: "alias"
        description: "FTP alias"
        required: true
        type: "string"
      - in: "query"
        name: "filePath"
        description: "File/folder location"
        required: true
        type: string
      produces:
      - 'application/json'
      - 'application/octet-stream'
      responses:
        "200":
          description: "File/folder retrieved successfuly"
        "400":
          description: "File doesn't exists"
        "401":
          description: "FlopBox bad authentication"
        "409":
          description: "Param error | Alias doesn't exists"